feat(am-dbg): add tx and log filtering
- add filters bar
- include log level in msgs
- split and reorg files
